Move page app specific middleware into the pages app
